29 erreurs, long long ok ?

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jsobel <jsobel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/01 16:08:21 by jsobel            #+#    #+#             */
-/*   Updated: 2018/11/14 19:13:30 by jsobel           ###   ########.fr       */
+/*   Updated: 2018/11/15 18:40:54 by jsobel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ typedef struct		s_data
 	va_list			arg;
 	int				nb;
 	long			nbl;
-	intmax_t		nbll;
+	long long		nbll;
 	char			c;
 	int				i;
 	int				count;
@@ -59,7 +59,7 @@ typedef struct		s_data
 
 int					ft_printf(const char *format, ...);
 
-char				*ft_itoa_intmax(intmax_t n);
+char				*ft_itoa_intmax(long long n);
 
 void				ft_printchar(t_data *ap);
